refactor: Move join screen and viewmodel to nickname package (#85)
The following files were renamed and moved:
- `JoinScreen.kt` -> `nickname/JoinNicknameScreen.kt`
- `JoinViewModel.kt` -> `nickname/JoinNicknameViewModel.kt`
